move to new file, configurable ttl, namespace changed cache validation
