+++ pkg/informer/informer_test.go
@@ -5,19 +5,14 @@ package informer
 
 import (
 	"context"
-	"github.com/stolostron/search-collector/pkg/config"
 	"testing"
 	"time"
 
-	"github.com/stolostron/search-v2-operator/api/v1alpha1"
-	"github.com/stretchr/testify/assert"
-	corev1 "k8s.io/api/core/v1"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/client-go/dynamic/fake"
-	fakeClient "k8s.io/client-go/kubernetes/fake"
 )
 
 // Create a GroupVersionResource
@@ -101,7 +96,7 @@ func Test_listAndResync(t *testing.T) {
 	informer, addFuncCount, _, _ := initInformer()
 
 	// Execute function
-	err := informer.listAndResync(nil)
+	err := informer.listAndResync()
 	if err != nil {
 		t.Error(err)
 	}
@@ -122,7 +117,7 @@ func Test_listAndResync_syncWithPrevState(t *testing.T) {
 	informer.resourceIndex["id-001"] = "some-resource-version"   // This resource won't get deleted.
 
 	// Execute function
-	err := informer.listAndResync(nil)
+	err := informer.listAndResync()
 	if err != nil {
 		t.Error(err)
 	}
@@ -248,7 +243,7 @@ func Test_watch(t *testing.T) {
 
 	// Start the watch() and wait until it is stopped.
 	go func() {
-		informer.watch(stopper, nil)
+		informer.watch(stopper)
 		close(done)
 	}()
 	// Wait 5 ms and send the signal to stop the watch()
@@ -274,341 +269,4 @@ func Test_WaitUntilInitialized_timeout(t *testing.T) {
 	if time.Since(start) > time.Duration(15)*time.Millisecond {
 		t.Errorf("Expected WaitUntilInitialized to time out within 15 milliseconds, but got %s", time.Since(start))
 	}
-}
-
-func TestFilterByGlobsIncludeNoExclude(t *testing.T) {
-	// Given: a list of namespaces: [foo, bar], include: [foo], exclude: []
-	namespaceList := &corev1.NamespaceList{
-		Items: []corev1.Namespace{
-			{ObjectMeta: v1.ObjectMeta{Name: "foo"}},
-			{ObjectMeta: v1.ObjectMeta{Name: "bar"}},
-		},
-	}
-
-	namespaceSelector := &v1alpha1.NamespaceSelector{
-		Include: []string{"foo"},
-		Exclude: []string{},
-	}
-
-	// When: we filter on namespaceList
-	result := filterByGlobs(namespaceList, namespaceSelector)
-
-	// Then: foo remains
-	assert.Equal(t, 1, len(result))
-	assert.Equal(t, true, result["foo"])
-}
-
-func TestFilterByGlobsExcludeNoInclude(t *testing.T) {
-	// Given: a list of namespaces: [foo, bar], include: [], exclude: [foo]
-	namespaceList := &corev1.NamespaceList{
-		Items: []corev1.Namespace{
-			{ObjectMeta: v1.ObjectMeta{Name: "foo"}},
-			{ObjectMeta: v1.ObjectMeta{Name: "bar"}},
-		},
-	}
-
-	namespaceSelector := &v1alpha1.NamespaceSelector{
-		Include: []string{},
-		Exclude: []string{"foo"},
-	}
-
-	// When: we filter on namespaceList
-	result := filterByGlobs(namespaceList, namespaceSelector)
-
-	// Then: bar remains
-	assert.Equal(t, 1, len(result))
-	assert.Equal(t, true, result["bar"])
-}
-
-func TestFilterByGlobsIncludeMatchExclude(t *testing.T) {
-	// Given: a list of namespaces: [foo, bar], include: [foo], exclude: [foo]
-	namespaceList := &corev1.NamespaceList{
-		Items: []corev1.Namespace{
-			{ObjectMeta: v1.ObjectMeta{Name: "foo"}},
-			{ObjectMeta: v1.ObjectMeta{Name: "bar"}},
-		},
-	}
-
-	namespaceSelector := &v1alpha1.NamespaceSelector{
-		Include: []string{"foo"},
-		Exclude: []string{"foo"},
-	}
-
-	// When: we filter on namespaceList
-	result := filterByGlobs(namespaceList, namespaceSelector)
-
-	// Then: no namespaces to collect
-	assert.Equal(t, 0, len(result))
-}
-
-func TestFilterByGlobsIncludeExclude(t *testing.T) {
-	// Given: a list of namespaces: [foo, bar], include: [foo], exclude: [bar]
-	namespaceList := &corev1.NamespaceList{
-		Items: []corev1.Namespace{
-			{ObjectMeta: v1.ObjectMeta{Name: "foo"}},
-			{ObjectMeta: v1.ObjectMeta{Name: "bar"}},
-		},
-	}
-
-	namespaceSelector := &v1alpha1.NamespaceSelector{
-		Include: []string{"foo"},
-		Exclude: []string{"bar"},
-	}
-
-	// When: we filter on namespaceList
-	result := filterByGlobs(namespaceList, namespaceSelector)
-
-	// Then: foo remains
-	assert.Equal(t, 1, len(result))
-	assert.Equal(t, true, result["foo"])
-}
-
-func TestFilterByGlobsIncludeWildcardNoExclude(t *testing.T) {
-	// Given: a list of namespaces: [foo, bar, baz], include: [ba*], exclude: []
-	namespaceList := &corev1.NamespaceList{
-		Items: []corev1.Namespace{
-			{ObjectMeta: v1.ObjectMeta{Name: "foo"}},
-			{ObjectMeta: v1.ObjectMeta{Name: "bar"}},
-			{ObjectMeta: v1.ObjectMeta{Name: "baz"}},
-		},
-	}
-
-	namespaceSelector := &v1alpha1.NamespaceSelector{
-		Include: []string{"ba*"},
-		Exclude: []string{},
-	}
-
-	// When: we filter on namespaceList
-	result := filterByGlobs(namespaceList, namespaceSelector)
-
-	// Then: bar and baz remain
-	assert.Equal(t, 2, len(result))
-	assert.Equal(t, true, result["bar"])
-	assert.Equal(t, true, result["baz"])
-}
-
-func TestFilterByGlobsExcludeWildcardNoInclude(t *testing.T) {
-	// Given: a list of namespaces: [foo, bar, baz], include: [], exclude: [ba*]
-	namespaceList := &corev1.NamespaceList{
-		Items: []corev1.Namespace{
-			{ObjectMeta: v1.ObjectMeta{Name: "foo"}},
-			{ObjectMeta: v1.ObjectMeta{Name: "bar"}},
-			{ObjectMeta: v1.ObjectMeta{Name: "baz"}},
-		},
-	}
-
-	namespaceSelector := &v1alpha1.NamespaceSelector{
-		Include: []string{},
-		Exclude: []string{"ba*"},
-	}
-
-	// When: we filter on namespaceList
-	result := filterByGlobs(namespaceList, namespaceSelector)
-
-	// Then: foo
-	assert.Equal(t, 1, len(result))
-	assert.Equal(t, true, result["foo"])
-}
-
-func TestFilterBySelectorsMatchLabels(t *testing.T) {
-	// Given a list of namespaces: [prod-app, dev-app, no-labels], selector matching env=prod
-	fc := fakeClient.NewSimpleClientset(
-		&corev1.Namespace{ObjectMeta: v1.ObjectMeta{Name: "prod-app", Labels: map[string]string{"env": "prod"}}},
-		&corev1.Namespace{ObjectMeta: v1.ObjectMeta{Name: "dev-app", Labels: map[string]string{"env": "dev"}}},
-		&corev1.Namespace{ObjectMeta: v1.ObjectMeta{Name: "no-labels"}},
-	)
-
-	nsSelector := &v1alpha1.NamespaceSelector{
-		MatchLabels: map[string]string{"env": "prod"},
-	}
-
-	// When: we filter by label selectors
-	result, err := filterBySelectors(fc, nsSelector)
-
-	// Then: prod-app remains
-	assert.NoError(t, err)
-	assert.Equal(t, 1, len(result.Items))
-	assert.Equal(t, "prod-app", result.Items[0].Name)
-}
-
-func TestFilterBySelectorsMatchExpressionsIn(t *testing.T) {
-	// Given a list of namespaces: [prod-app, dev-app, no-labels], selector matching env=prod
-	fc := fakeClient.NewSimpleClientset(
-		&corev1.Namespace{ObjectMeta: v1.ObjectMeta{Name: "prod-app", Labels: map[string]string{"env": "prod"}}},
-		&corev1.Namespace{ObjectMeta: v1.ObjectMeta{Name: "dev-app", Labels: map[string]string{"env": "dev"}}},
-		&corev1.Namespace{ObjectMeta: v1.ObjectMeta{Name: "no-labels"}},
-	)
-
-	nsSelector := &v1alpha1.NamespaceSelector{
-		MatchExpressions: []v1.LabelSelectorRequirement{
-			{Key: "env", Values: []string{"prod"}, Operator: v1.LabelSelectorOpIn},
-		},
-	}
-
-	// When: we filter by expression selectors
-	result, err := filterBySelectors(fc, nsSelector)
-
-	// Then: prod-app remains
-	assert.NoError(t, err)
-	assert.Equal(t, 1, len(result.Items))
-	assert.Equal(t, "prod-app", result.Items[0].Name)
-}
-
-func TestFilterBySelectorsMatchExpressionsNotIn(t *testing.T) {
-	// Given a list of namespaces: [prod-app, dev-app, no-labels], selector matching env!=prod
-	fc := fakeClient.NewSimpleClientset(
-		&corev1.Namespace{ObjectMeta: v1.ObjectMeta{Name: "prod-app", Labels: map[string]string{"env": "prod"}}},
-		&corev1.Namespace{ObjectMeta: v1.ObjectMeta{Name: "dev-app", Labels: map[string]string{"env": "dev"}}},
-		&corev1.Namespace{ObjectMeta: v1.ObjectMeta{Name: "no-labels"}},
-	)
-
-	nsSelector := &v1alpha1.NamespaceSelector{
-		MatchExpressions: []v1.LabelSelectorRequirement{
-			{Key: "env", Values: []string{"prod"}, Operator: v1.LabelSelectorOpNotIn},
-		},
-	}
-
-	// When: we filter by expression selectors
-	result, err := filterBySelectors(fc, nsSelector)
-
-	// Then: dev-app and no-labels remains
-	assert.NoError(t, err)
-	assert.Equal(t, 2, len(result.Items))
-	assert.Equal(t, "dev-app", result.Items[0].Name)
-	assert.Equal(t, "no-labels", result.Items[1].Name)
-}
-
-func TestFilterBySelectorsMatchExpressionsNotInWithLabels(t *testing.T) {
-	// Given a list of namespaces: [prod-app, dev-app, no-labels], selector matching env!=prod and label env=dev
-	fc := fakeClient.NewSimpleClientset(
-		&corev1.Namespace{ObjectMeta: v1.ObjectMeta{Name: "prod-app", Labels: map[string]string{"env": "prod"}}},
-		&corev1.Namespace{ObjectMeta: v1.ObjectMeta{Name: "dev-app", Labels: map[string]string{"env": "dev"}}},
-		&corev1.Namespace{ObjectMeta: v1.ObjectMeta{Name: "no-labels"}},
-	)
-
-	nsSelector := &v1alpha1.NamespaceSelector{
-		MatchExpressions: []v1.LabelSelectorRequirement{
-			{Key: "env", Values: []string{"prod"}, Operator: v1.LabelSelectorOpNotIn},
-		},
-		MatchLabels: map[string]string{"env": "dev"},
-	}
-
-	// When: we filter by expression selectors
-	result, err := filterBySelectors(fc, nsSelector)
-
-	// Then: dev-app remains
-	assert.NoError(t, err)
-	assert.Equal(t, 1, len(result.Items))
-	assert.Equal(t, "dev-app", result.Items[0].Name)
-}
-
-func TestFilterBySelectorsThenGlobs(t *testing.T) {
-	// Given a list of namespaces: [prod-app, dev-app, dev-app-2, no-labels], selector matching env!=prod and label env=dev and glob Include:[dev-*]
-	fc := fakeClient.NewSimpleClientset(
-		&corev1.Namespace{ObjectMeta: v1.ObjectMeta{Name: "prod-app", Labels: map[string]string{"env": "prod"}}},
-		&corev1.Namespace{ObjectMeta: v1.ObjectMeta{Name: "dev-app", Labels: map[string]string{"env": "dev"}}},
-		&corev1.Namespace{ObjectMeta: v1.ObjectMeta{Name: "dev-app-2", Labels: map[string]string{"env": "dev"}}},
-		&corev1.Namespace{ObjectMeta: v1.ObjectMeta{Name: "no-labels"}},
-	)
-
-	nsSelector := &v1alpha1.NamespaceSelector{
-		MatchExpressions: []v1.LabelSelectorRequirement{
-			{Key: "env", Values: []string{"prod"}, Operator: v1.LabelSelectorOpNotIn},
-		},
-		MatchLabels: map[string]string{"env": "dev"},
-		Include:     []string{"dev-*"},
-		Exclude:     []string{},
-	}
-
-	// When: we filter by expression selectors and glob
-	interim, err := filterBySelectors(fc, nsSelector)
-	result := filterByGlobs(interim, nsSelector)
-
-	// Then: dev-app and dev-app-2 remain
-	assert.NoError(t, err)
-	assert.Equal(t, 2, len(result))
-	assert.Equal(t, true, result["dev-app"])
-	assert.Equal(t, true, result["dev-app-2"])
-}
-
-func TestFilterByEmptySelectors(t *testing.T) {
-	// Given: a list of namespaces: [foo, bar, baz] with no filters
-	fc := fakeClient.NewSimpleClientset(
-		&corev1.Namespace{ObjectMeta: v1.ObjectMeta{Name: "foo"}},
-		&corev1.Namespace{ObjectMeta: v1.ObjectMeta{Name: "bar"}},
-		&corev1.Namespace{ObjectMeta: v1.ObjectMeta{Name: "baz"}},
-	)
-	namespaceSelector := &v1alpha1.NamespaceSelector{}
-
-	// When: we filter on namespaceList
-	interim, err := filterBySelectors(fc, namespaceSelector)
-	result := filterByGlobs(interim, namespaceSelector)
-
-	// Then: foo, bar, and baz remain
-	assert.NoError(t, err)
-	assert.Equal(t, 3, len(result))
-	assert.Equal(t, true, result["foo"])
-	assert.Equal(t, true, result["bar"])
-	assert.Equal(t, true, result["baz"])
-}
-
-// Feature flag disabled: allow everything regardless of map contents.
-func TestIsNamespaceAllowed_FeatureDisabled(t *testing.T) {
-	original := config.Cfg.FeatureConfigurableCollection
-	defer func() { config.Cfg.FeatureConfigurableCollection = original }()
-	config.Cfg.FeatureConfigurableCollection = false
-
-	assert.True(t, isNamespaceAllowed(nil, "any-ns"))
-	assert.True(t, isNamespaceAllowed(map[string]bool{}, "any-ns"))
-	assert.True(t, isNamespaceAllowed(map[string]bool{"other": true}, "any-ns"))
-}
-
-// Nil map means no filter was configured: allow all namespaces.
-func TestIsNamespaceAllowed_NilMap(t *testing.T) {
-	original := config.Cfg.FeatureConfigurableCollection
-	defer func() { config.Cfg.FeatureConfigurableCollection = original }()
-	config.Cfg.FeatureConfigurableCollection = true
-
-	assert.True(t, isNamespaceAllowed(nil, "any-ns"))
-}
-
-// Empty map means selector matched zero namespaces: block everything.
-func TestIsNamespaceAllowed_EmptyMap(t *testing.T) {
-	original := config.Cfg.FeatureConfigurableCollection
-	defer func() { config.Cfg.FeatureConfigurableCollection = original }()
-	config.Cfg.FeatureConfigurableCollection = true
-
-	assert.False(t, isNamespaceAllowed(map[string]bool{}, "any-ns"))
-}
-
-// Cluster-scoped resources (empty namespace) always pass.
-func TestIsNamespaceAllowed_ClusterScoped(t *testing.T) {
-	original := config.Cfg.FeatureConfigurableCollection
-	defer func() { config.Cfg.FeatureConfigurableCollection = original }()
-	config.Cfg.FeatureConfigurableCollection = true
-
-	assert.True(t, isNamespaceAllowed(map[string]bool{"ns1": true}, ""))
-	assert.True(t, isNamespaceAllowed(map[string]bool{}, ""))
-}
-
-// Namespace in the allowed map passes.
-func TestIsNamespaceAllowed_Allowed(t *testing.T) {
-	original := config.Cfg.FeatureConfigurableCollection
-	defer func() { config.Cfg.FeatureConfigurableCollection = original }()
-	config.Cfg.FeatureConfigurableCollection = true
-
-	allowed := map[string]bool{"ns1": true, "ns2": true}
-	assert.True(t, isNamespaceAllowed(allowed, "ns1"))
-	assert.True(t, isNamespaceAllowed(allowed, "ns2"))
-}
-
-// Namespace not in the allowed map is blocked.
-func TestIsNamespaceAllowed_NotAllowed(t *testing.T) {
-	original := config.Cfg.FeatureConfigurableCollection
-	defer func() { config.Cfg.FeatureConfigurableCollection = original }()
-	config.Cfg.FeatureConfigurableCollection = true
-
-	allowed := map[string]bool{"ns1": true}
-	assert.False(t, isNamespaceAllowed(allowed, "ns2"))
-	assert.False(t, isNamespaceAllowed(allowed, "kube-system"))
 }
